pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cmc-csci040-brycekage-pythonllm"
version = "0.1.3"
requires-python = ">=3.11"
description = "A terminal LLM chat REPL powered by Groq."
readme = "README.md"
dependencies = [
    "groq",
    "python-dotenv",
    "flake8",
    "gitpython",
]

[project.scripts]
chat = "chat:repl"

[tool.hatch.build.targets.wheel]
include = ["chat.py", "tools/"]

[tool.pytest.ini_options]
addopts = "--ignore=test_projects"